pyproject.toml
[project]
name = "role-play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.117.1",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.36.0",
]

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple"
default = true

[dependency-groups]
dev = ["pytest"]
lint = ["mypy", "ruff", "types-passlib"]

[tool.mypy]
mypy_path = ["src"]
